requirements.txt
fastapi==0.99.1
pydantic==1.10.11
uvicorn==0.22.0
gunicorn==20.1.0
pymongo==4.4.1
python-dotenv==1.0.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1
PyJWT==1.7.1
fastapi-jwt-auth[asymmetric]==0.5.0
email-validator==2.0.0.post2
beanie==1.21.0
motor==3.2.0
# Dev dependencies
black==23.3.0
isort==5.12.0
autoflake==2.2.0
flake8==6.0.0
pytest==7.4.0